@@ -6,33 +6,41 @@
   "author": "Frank Wei",
   "license": "MIT",
   "scripts": {
-    "start": "ts-node ./src/app.ts",
+    "start": "ts-node -r tsconfig-paths/register ./src/app.ts",
     "lint": "eslint",
-    "dev": "ts-node-dev --respawn --transpileOnly ./src/app.ts",
-    "typeorm": "ts-node -r tsconfig-paths/register ./node_modules/typeorm/cli.js"
+    "dev": "ts-node-dev --respawn --transpileOnly -r tsconfig-paths/register ./src/app.ts",
+    "typeorm": "ts-node -r tsconfig-paths/register ./node_modules/typeorm/cli.js",
+    "test": "jest"
   },
   "dependencies": {
-    "@types/eth-sig-util": "^2.1.0",
-    "@types/node": "^12.12.17",
     "axios": "^0.19.0",
+    "cron": "^1.7.2",
     "dotenv": "^8.2.0",
     "eth-sig-util": "^2.5.2",
     "inversify": "^5.0.1",
-    "mysql2": "^2.0.2",
-    "node-cron": "^2.0.3",
-    "pg": "^7.14.0",
+    "log4js": "^6.1.0",
+    "mysql2": "^2.1.0",
+    "pg": "^7.17.0",
     "reflect-metadata": "^0.1.13",
+    "table": "^5.4.6",
     "telegraf": "^3.34.1",
-    "tsconfig-paths": "^3.9.0",
-    "typeorm": "^0.2.21",
+    "typeorm": "^0.2.22",
     "web3": "^1.2.4"
   },
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^2.11.0",
-    "@typescript-eslint/parser": "^2.11.0",
-    "eslint": "^6.7.2",
+    "@types/cron": "^1.7.1",
+    "@types/eth-sig-util": "^2.1.0",
+    "@types/jest": "^24.0.25",
+    "@types/node": "^13.1.1",
+    "@types/table": "^4.0.7",
+    "@typescript-eslint/eslint-plugin": "^2.13.0",
+    "@typescript-eslint/parser": "^2.13.0",
+    "eslint": "^6.8.0",
+    "jest": "^24.9.0",
+    "ts-jest": "^24.2.0",
     "ts-node": "^8.5.4",
     "ts-node-dev": "^1.0.0-pre.44",
-    "typescript": "^3.7.3"
+    "tsconfig-paths": "^3.9.0",
+    "typescript": "^3.7.4"
   }
 }
