@@ -1,5 +1,13 @@
 export const Injections = {
+    Container: Symbol.for("Container"),
     Controller: Symbol.for("Controller"),
-    BotService: Symbol.for("BotService"),
     Context: Symbol.for("Context"),
+    Repository: Symbol.for("Repository"),
+    Scheduler: Symbol.for("Scheduler"),
+
+    BotService: Symbol.for("BotService"),
+    DatabaseService: Symbol.for("DatabaseService"),
+    LoggerService: Symbol.for("LoggerService"),
+    MatatakiService: Symbol.for("MatatakiService"),
+    Web3Service: Symbol.for("Web3Service"),
 };
