@@ -1,5 +1,7 @@
-import { Controller, Command } from "../decorators";
-import { MessageHandlerContext } from "../definitions";
+import { inject } from "inversify";
+
+import { Controller, Command } from "#/decorators";
+import { MessageHandlerContext } from "#/definitions";
 import { BaseController } from ".";
 
 @Controller("debug")
@@ -17,5 +19,16 @@ export class DebugController extends BaseController<DebugController> {
     @Command("throwerrorasync")
     throwErrorAsync() {
         return Promise.reject(new Error("An intentional error"));
+    }
+
+    @Command("table")
+    displayTable({ replyWithHTML }: MessageHandlerContext) {
+        return replyWithHTML(`<pre>
+| Tables   |      Are      |  Cool |
+|----------|:-------------:|------:|
+| col 1 is |  left-aligned | $1600 |
+| col 2 is |    centered   |   $12 |
+| col 3 is | right-aligned |    $1 |
+</pre>`);
     }
 }
