@@ -1,4 +1,4 @@
-export type CommandDefinition = {
+export type CommandHandlerInfo = {
     name: string,
     methodName: string,
     ignorePrefix: boolean,
