@@ -1,8 +1,8 @@
 import { ContextMessageUpdate } from "telegraf";
-import { IncomingMessage } from "telegraf/typings/telegram-types";
+import { IncomingMessage, User } from "telegraf/typings/telegram-types";
 
 export type MessageHandlerContext = Omit<ContextMessageUpdate, "message"> & {
-    message: Omit<IncomingMessage, "text"> & { text: string };
+    message: Omit<IncomingMessage, "text"> & { text: string } & Omit<IncomingMessage, "from"> & { from: User };
 }
 
 export interface MessageHandler {
