@@ -1,2 +1,5 @@
-export * from "./CommandDefinition";
+export * from "./AssociatedInfo";
+export * from "./CommandHandlerInfo";
+export * from "./EventHandlerInfo";
+export * from "./GroupRequirement";
 export * from "./MessageHandler";
