@@ -1,11 +1,28 @@
-import { Entity, PrimaryColumn, ManyToMany } from "typeorm";
+import { Entity, PrimaryColumn, ManyToMany, Column, Index } from "typeorm";
 
 import { User } from "./User";
+import { GroupRequirement } from "definitions/GroupRequirement";
 
 @Entity()
 export class Group {
-    @PrimaryColumn()
-    id!: number;
+    @PrimaryColumn({ type: "bigint" })
+    id!: number | string;
+
+    @Column({ type: "text" })
+    title!: string;
+
+    @Column({ type: "bigint" })
+    @Index()
+    creatorId!: number | string;
+
+    @Column({ type: "boolean", default: true })
+    active!: boolean;
+
+    @Column({ type: "int" })
+    tokenId!: number;
+
+    @Column({ type: "jsonb" })
+    requirement!: GroupRequirement;
 
     @ManyToMany(type => User, user => user.groups)
     members!: User[];
