@@ -1,15 +1,11 @@
-import { Entity, PrimaryColumn, OneToMany, ManyToMany, JoinTable } from "typeorm";
+import { Entity, PrimaryColumn, ManyToMany, JoinTable } from "typeorm";
 
-import { Wallet } from "./Wallet";
 import { Group } from "./Group";
 
 @Entity()
 export class User {
-    @PrimaryColumn()
-    id!: number;
-
-    @OneToMany(type => Wallet, wallet => wallet.user)
-    wallets!: Wallet[];
+    @PrimaryColumn({ type: "bigint" })
+    id!: number | string;
 
     @ManyToMany(type => Group, group => group.members)
     @JoinTable({ name: "group_member" })
