@@ -1,4 +1,12 @@
-export * from "./Group";
+import { Group } from "./Group";
+import { User } from "./User";
 
-export * from "./User";
-export * from "./Wallet";
+export {
+    Group,
+    User,
+};
+
+export const entities = [
+    Group,
+    User,
+];
