@@ -1,9 +1,5 @@
-import { BotService } from "./BotService";
-
-export {
-    BotService,
-};
-
-export const services = [
-    BotService,
-];
+export * from "./IBotService";
+export * from "./IDatabaseService";
+export * from "./ILoggerService";
+export * from "./IMatatakiService";
+export * from "./IWeb3Service";
